requirements.txt
Django==3.2
djangorestframework==3.12.4
django-mysql==3.11.1
mysqlclient==2.0.3
requests==2.25.1